chore: remove .nvmrc
To reduce clutter in the root dir, .nvmrc can be removed.  Instead of
using this file to indicate the version that we prefer, the CONTRIBUTING.md
now specifies the version.

In the near future, we should bump this version to 22.x, because 20.x recently
went into maintenance (22 is active LTS).  In addition, with more tests we
should go to full support for LTS versions, but for now we are focusing on
this one which is the most tested.

Part of #21

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>