docs: add role and tools to getting started cli experience
Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

Fixes #148

We have found that bees perform best when:
* they use fewer number of tools, and
* they are provided explicit instructions/role

Here we updated the getting started section of the README to
incorporate those recommendations into the simple example. We update
the system prompt to provide explicit instructions, and we remove the
search tool as it isn't needed for weather-related queries. It also
adds some tips related to these two points to the appropriate section.

Additionally, it adds a yarn helper method to run the simple example:
`yarn run quickstart`

Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>